long overdue commit for AquariWM state + async X11 handling
I can't get the async stuff to work due to lifetime things I don't understand. I also don't know if it is even the best thing to do because of the fact XCB doesn't provide any way to use callbacks to know when events or replies have been received, which is pretty much the async standard nowadays and is a lot better than constantly polling.